Rename KBaseIdentifierFactory -> IdentifierUtils
